[♻️refactor]: 잘못된 import 구문 수정
